create a separate component for save button
